--- a/vllm_ascend/ascend_config.py
+++ b/vllm_ascend/ascend_config.py
@@ -62,6 +62,10 @@ class AscendConfig:
                     "lmhead_tensor_parallel_size is only supported in the pure DP scenario"
                 )
 
+        self.enable_mla_sp = additional_config.get("enable_mla_sp", False)
+        self.o_shard_parallel_size = int(additional_config.get("o_shard_parallel_size", -1))
+        self.enable_o_shard = self.o_shard_parallel_size > 0
+        self.o_shard_full_layers = int(additional_config.get("o_shard_full_layers", 0))
 
 class TorchairGraphConfig:
     """