### What this PR does / why we need it?
The DeepSeekV2 MLA computation is facing two problems under multi-way DP configuration:
- Incoming requests often have varying input sequence lengths, leading to the unbalance among each DP.
- When there aren't many requests, some DP shards will just do dummy run and become useless.

This PR introduces a staged hybrid parallelism strategy optimized for MLA computation during the prefill phase:
<div align="center">

![pic](https://github.com/user-attachments/assets/8de03232-191c-4ba2-9619-fa2c7b2d8aa6)

Picture from https://arxiv.org/pdf/2506.12708 Fig. 16
</div>

- Stage 1: All the input tokens among all DP shards will be evenly rearranged to each rank. (SP)
- Stage 2: For each DP, All-Gather to get its own tokens and apply TP to run attention. After solving the KV Cache problem, this stage will be further optimized such that each DP is able to handle all the global tokens instead.
- Stage 3: All-to-All to run SP for `o_proj`. Here the weight of `o_proj` matrix will be completed. To save the memory, this PR introduces a shard approach. The weights of `o_proj` are kept at different ranks, and broadcast asynchronously to get the complete matrix in `FRACTAL_NZ` format for current layer during the computation.

### Does this PR introduce _any_ user-facing change?
This PR introduces three new configs in `additional_config`:
| Name | Effect | Required | Type | Constraints |
|---------|---------|---------|---------|---------|
| `enable_mla_sp` | Whether to enable SP for DeepSeekV2 MLA. | No | bool | Must be enabled with `o_proj` shard feature. It is temporary disabled when some other features are enabled. |
| `o_shard_parallel_size` | How to divide the rows of the `o_proj` matrix evenly across ranks for shard. | No | int | It should be at least 2 and should be a divisor of `world_size`. When it is missing, the shard feature is disabled. |
| `o_shard_full_layers` | The first complete `o_shard_full_layers` layers will be kept among all ranks without shard. When computing the `k`-th layer, the `o_proj` weight of the `k + o_shard_full_layers`-th layer will be broadcast asynchronously. | No | int | Default value is `0`. |

Example:
`--additional-config '{"enable_mla_sp": true, "o_shard_parallel_size": 8, "o_shard_full_layers": 5}'`

### How was this patch tested?
vLLM main: https://github.com/vllm-project/vllm/commit/a4454e9401c9c76150dc3a4d23d10377926b2fb6

- For DP 4 TP 4 configuration, start script:
```
export VLLM_USE_V1=1
export TASK_QUEUE_ENABLE=1

python -m vllm.entrypoints.openai.api_server --model=DeepSeek-R1-W8A8-VLLM \
 --served-model-name auto \
 --load-format=auto \
 --trust-remote-code \
 --enforce_eager \
 --distributed-executor-backend=mp \
 -tp=4 \
 -dp=4 \
 --quantization ascend \
 --additional-config '{"chunked_prefill_for_mla": true, "enable_mla_sp": true, "o_shard_parallel_size": 8, "o_shard_full_layers": 5}' \
 --port 8006 \
 --max-num-seqs 24 \
 --max-model-len 8192 \
 --max-num-batched-tokens 8192 \
 --block-size 128 \
 --gpu-memory-utilization 0.96 \
 --no-enable-prefix-caching
```
| max-concurrency | #requests | input tokens distribution | SP Throughput | Baseline Throughput |
|---------|---------|---------|---------|---------|
| 1 | 200 | Fixed 4096 | **0.77** | 0.64 |
| 96 | 400 | Fixed 4096 | **1.87** | 1.30 |
| 96 | 400 | Randomly from [1 .. 4096] | **3.67** | 2.44 |

- For DP 1 TP 16 configuration, the `max-model-len` and `max-num-batched-tokens` are `32768`.

| max-concurrency | #requests | input tokens distribution | SP Throughput | Baseline Throughput |
|---------|---------|---------|---------|---------|
| 1 | 200 | Fixed 4096 | 1.35 | **1.52** |
| 24 | 200 | Fixed 4096 | **2.29** | 2.21 |

- vLLM version: v0.10.1.1
- vLLM main: https://github.com/vllm-project/vllm/commit/c07a73317d202c2dad67f12893fcddb6d3664950