@weijinqian0 2025-08-27
The solution may not be sufficiently versatile.